Changing models, repositories and services to be easier to read. Deleting menu and util package.

+++ src/main/java/com/example/GuesthouseWebpage/repository/BookingPaymentRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookingPaymentRepository extends JpaRepository<BookingPayment, Integer> {
+public interface BookingPaymentRepository extends JpaRepository<BookingPayment, Long> {
+
 }
