+++ src/main/java/com/example/GuesthouseWebpage/repository/MealTypeRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MealTypeRepository extends JpaRepository<MealType, Integer> {
+public interface MealTypeRepository extends JpaRepository<MealType, Long> {
+
 
 }
