+++ src/main/java/com/example/GuesthouseWebpage/repository/RoomTypeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoomTypeRepository extends JpaRepository<RoomType, Integer> {
+public interface RoomTypeRepository extends JpaRepository<RoomType, Long> {
 
 }
